Created a Portable Text Component to render content.
If you refresh your front end for this post now, you won't see the image rendered on the page. This is because the PortableText component does not know what to do with it. To solve this, you can create an object of components that can replace or extend the defaults.